@@ -9,31 +9,31 @@
     <version>0.1.0-SNAPSHOT</version>
 
     <properties>
-        <java.version>13</java.version>
+        <java.version>17</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 
-        <assertj-core.version>3.17.1</assertj-core.version>
-        <azure-identity.version>1.6.0</azure-identity.version>
-        <commons-lang3.version>3.11</commons-lang3.version>
+        <assertj-core.version>3.26.3</assertj-core.version>
+        <azure-identity.version>1.13.3</azure-identity.version>
+        <commons-lang3.version>3.17.0</commons-lang3.version>
         <commons-math3.version>3.6.1</commons-math3.version>
-        <flyway.version>6.5.5</flyway.version>
-        <google-api-client.version>1.30.10</google-api-client.version>
-        <guava.version>32.0.0-jre</guava.version>
-        <jacoco.version>0.8.4</jacoco.version>
-        <jackson-databind.version>2.14.0-rc2</jackson-databind.version>
-        <jjwt.version>0.11.2</jjwt.version>
-        <junit-jupiter.version>5.1.0</junit-jupiter.version>
-        <log4j.version>2.17.1</log4j.version>
-        <lombok.version>1.18.16</lombok.version>
-        <microsoft-graph.version>5.38.0</microsoft-graph.version>
-        <mockito.version>3.2.4</mockito.version>
-        <vertx.version>4.3.8</vertx.version>
-        <postgresql.version>42.3.9</postgresql.version>
+        <flyway.version>10.18.1</flyway.version>
+        <google-api-client.version>2.7.0</google-api-client.version>
+        <guava.version>33.3.1-jre</guava.version>
+        <jacoco.version>0.8.12</jacoco.version>
+        <jackson-databind.version>2.17.2</jackson-databind.version>
+        <jjwt.version>0.12.6</jjwt.version>
+        <junit-jupiter.version>5.11.0</junit-jupiter.version>
+        <log4j.version>2.24.0</log4j.version>
+        <lombok.version>1.18.34</lombok.version>
+        <microsoft-graph.version>6.16.0</microsoft-graph.version>
+        <mockito.version>5.13.0</mockito.version>
+        <vertx.version>4.5.10</vertx.version>
+        <postgresql.version>42.7.4</postgresql.version>
         <sentry-log4j2.version>1.7.30</sentry-log4j2.version>
 
-        <maven-jar-plugin.version>3.0.0</maven-jar-plugin.version>
-        <maven-shade-plugin.version>3.2.1</maven-shade-plugin.version>
-        <maven-surefire-plugin.version>2.22.2</maven-surefire-plugin.version>
+        <maven-jar-plugin.version>3.4.2</maven-jar-plugin.version>
+        <maven-shade-plugin.version>3.6.0</maven-shade-plugin.version>
+        <maven-surefire-plugin.version>3.5.0</maven-surefire-plugin.version>
     </properties>
 
     <dependencies>
@@ -46,6 +46,12 @@
             <groupId>org.flywaydb</groupId>
             <artifactId>flyway-core</artifactId>
             <version>${flyway.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.flywaydb</groupId>
+            <artifactId>flyway-database-postgresql</artifactId>
+            <version>${flyway.version}</version>
+            <scope>runtime</scope>
         </dependency>
 
         <dependency>
@@ -192,12 +198,6 @@
         <dependency>
             <groupId>org.mockito</groupId>
             <artifactId>mockito-core</artifactId>
-            <version>${mockito.version}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.mockito</groupId>
-            <artifactId>mockito-inline</artifactId>
             <version>${mockito.version}</version>
             <scope>test</scope>
         </dependency>
