@@ -1,11 +1,5 @@
 package nl.cofx.top10.quiz;
 
-import static nl.cofx.top10.postgresql.PostgreSql.toUuid;
-
-import java.time.Instant;
-import java.util.*;
-import java.util.stream.Collectors;
-
 import io.vertx.core.Future;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
@@ -14,6 +8,14 @@ import lombok.extern.log4j.Log4j2;
 import nl.cofx.top10.ConflictException;
 import nl.cofx.top10.NotFoundException;
 import nl.cofx.top10.quiz.dto.*;
+
+import java.time.Instant;
+import java.time.OffsetDateTime;
+import java.time.ZoneOffset;
+import java.util.*;
+import java.util.stream.Collectors;
+
+import static nl.cofx.top10.postgresql.PostgreSql.toUuid;
 
 @Log4j2
 public class QuizRepository {
@@ -95,7 +97,7 @@ public class QuizRepository {
                 .isActive(array.getBoolean(2))
                 .creatorId(creatorId)
                 .isCreator(creatorId.equals(accountId))
-                .deadline(array.getInstant(4));
+                .deadline(((OffsetDateTime) array.getValue(4)).toInstant());
 
         var personalListId = array.getString(5);
         if (personalListId != null) {
@@ -195,7 +197,7 @@ public class QuizRepository {
 
     public Future<String> createQuiz(SQLConnection connection, String name, String creatorId, Instant deadline) {
         return Future.future(promise -> {
-            var params = new JsonArray().add(name).add(toUuid(creatorId)).add(deadline);
+            var params = new JsonArray().add(name).add(toUuid(creatorId)).add(deadline.atOffset(ZoneOffset.UTC));
             connection.updateWithParams(CREATE_QUIZ_TEMPLATE, params, asyncResult -> {
                 if (asyncResult.failed()) {
                     var cause = asyncResult.cause();
