@@ -1,24 +1,24 @@
 package nl.cofx.top10.session;
 
-import static nl.cofx.top10.session.SessionConfiguration.JWT_COOKIE_NAME;
-import static nl.cofx.top10.session.SessionConfiguration.SESSION_EXPIRATION_IN_SECONDS;
-
-import java.time.Instant;
-import java.util.Date;
-
-import javax.crypto.SecretKey;
-
 import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
 import io.vertx.core.AbstractVerticle;
 import io.vertx.core.buffer.Buffer;
-import io.vertx.core.http.*;
+import io.vertx.core.http.Cookie;
+import io.vertx.core.http.CookieSameSite;
+import io.vertx.core.http.HttpMethod;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.web.Router;
 import io.vertx.ext.web.RoutingContext;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import nl.cofx.top10.jwt.Jwt;
+
+import javax.crypto.SecretKey;
+import java.time.Instant;
+import java.util.Date;
+
+import static nl.cofx.top10.session.SessionConfiguration.JWT_COOKIE_NAME;
+import static nl.cofx.top10.session.SessionConfiguration.SESSION_EXPIRATION_IN_SECONDS;
 
 @Log4j2
 @RequiredArgsConstructor
@@ -50,7 +50,7 @@ public class SessionStatusVerticle extends AbstractVerticle {
         var response = routingContext.response();
         response.putHeader("content-type", "application/json");
 
-        var existingCookie = routingContext.getCookie(JWT_COOKIE_NAME);
+        var existingCookie = routingContext.request().getCookie(JWT_COOKIE_NAME);
         if (existingCookie == null) {
             log.debug("No session cookie present");
             response.end(NO_SESSION_RESPONSE);
@@ -66,16 +66,16 @@ public class SessionStatusVerticle extends AbstractVerticle {
 
         log.debug("Extending expiration date of session cookie");
 
-        var body = jws.getBody();
+        var body = jws.getPayload();
         var subject = body.getSubject();
         var name = (String) body.get("name");
         var emailAddress = (String) body.get("emailAddress");
         var jwt = Jwts.builder()
-                .setExpiration(Date.from(Instant.now().plusSeconds(SESSION_EXPIRATION_IN_SECONDS)))
-                .setSubject(subject)
+                .expiration(Date.from(Instant.now().plusSeconds(SESSION_EXPIRATION_IN_SECONDS)))
+                .subject(subject)
                 .claim("name", name)
                 .claim("emailAddress", emailAddress)
-                .signWith(secretKey, SignatureAlgorithm.HS512)
+                .signWith(secretKey, Jwts.SIG.HS512)
                 .compact();
 
         var newCookie = Cookie.cookie(JWT_COOKIE_NAME, jwt)
