@@ -1,17 +1,14 @@
 package nl.cofx.top10.jwt;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-import java.util.Date;
-
-import javax.crypto.SecretKey;
-
-import org.junit.jupiter.api.Test;
-
 import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
 import io.jsonwebtoken.io.Decoders;
 import io.jsonwebtoken.security.Keys;
+import org.junit.jupiter.api.Test;
+
+import javax.crypto.SecretKey;
+import java.util.Date;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 class JwtTest {
 
@@ -24,13 +21,13 @@ class JwtTest {
     public void convertsValidJwt() {
         var subject = "test";
         var validJwt = Jwts.builder()
-                .setSubject(subject)
-                .signWith(SECRET_KEY, SignatureAlgorithm.HS512)
+                .subject(subject)
+                .signWith(SECRET_KEY, Jwts.SIG.HS512)
                 .compact();
 
         var jws = jwt.getJws(validJwt);
         assertThat(jws).isNotNull();
-        assertThat(jws.getBody().getSubject()).isEqualTo(subject);
+        assertThat(jws.getPayload().getSubject()).isEqualTo(subject);
     }
 
     @Test
@@ -46,9 +43,9 @@ class JwtTest {
     @Test
     public void returnsNullGivenExpiredJwt() {
         var validJwt = Jwts.builder()
-                .setExpiration(new Date())
-                .setSubject("test")
-                .signWith(SECRET_KEY, SignatureAlgorithm.HS512)
+                .expiration(new Date())
+                .subject("test")
+                .signWith(SECRET_KEY, Jwts.SIG.HS512)
                 .compact();
 
         assertThat(jwt.getJws(validJwt)).isNull();
